pyproject.toml
[project]
name = "html2md"
version = "0.1.0"
description = "把网页文章一键转成 VuePress 2 Markdown 的转载工具"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MPL-2.0" }
dependencies = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "readability-lxml>=0.8.1",
    "markdownify>=0.13",
    "typer>=0.12",
    "python-slugify>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
html2md = "html2md.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/html2md"]
